Rename grouping module to groups and update other modules and test.
