+++ pennylane/groups/graph_colouring.py
