+++ pennylane/groups/pauli.py
@@ -12,14 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Functions for generating and manipulating elements of the Pauli group.
+Functions for constructing the N-qubit Pauli group, and performing the 
+group operation (multiplication).
 """
 
 import numpy as np
 
 from pennylane import Identity
-from pennylane.grouping.utils import binary_to_pauli, pauli_to_binary, are_identical_pauli_words
-
+from pennylane.groups.pauli_utils import (
+    binary_to_pauli,
+    pauli_to_binary,
+    are_identical_pauli_words
+)
 
 def pauli_group(n_qubits, wire_map=None):
     """Generator for iterating over the n-qubit Pauli group.
@@ -48,7 +52,7 @@ def pauli_group(n_qubits, wire_map=None):
 
     .. code-block:: python
 
-        from pennylane.grouping.pauli_group import pauli_group
+        from pennylane.groups.pauli import pauli_group
 
         n_qubits = 3
 
@@ -114,7 +118,7 @@ def pauli_mult(pauli_1, pauli_2, wire_map=None):
     This function enables multiplication of Pauli group elements at the level of
     Pauli words, rather than matrices. For example,
 
-    >>> from pennylane.grouping.pauli_group import pauli_mult
+    >>> from pennylane.groups.pauli import pauli_mult
     >>> pauli_1 = qml.PauliX(0) @ qml.PauliZ(1)
     >>> pauli_2 = qml.PauliY(0) @ qml.PauliZ(1)
     >>> product = pauli_mult(pauli_1, pauli_2)
@@ -172,7 +176,7 @@ def pauli_mult_with_phase(pauli_1, pauli_2, wire_map=None):
     phase accumulated as a result of the Pauli product rules
     :math:`\sigma_i \sigma_j = i \sigma_k`.
 
-    >>> from pennylane.grouping.pauli_group import pauli_mult_with_phase
+    >>> from pennylane.groups.pauli import pauli_mult_with_phase
     >>> pauli_1 = qml.PauliX(0) @ qml.PauliZ(1)
     >>> pauli_2 = qml.PauliY(0) @ qml.PauliZ(1)
     >>> product, phase = pauli_mult_with_phase(pauli_1, pauli_2)
