+++ pennylane/tape/tape.py
@@ -141,7 +141,7 @@ def expand_tape(tape, depth=1, stop_at=None, expand_measurements=False):
     # expansion acts on the original tape in place.
     if tape._obs_sharing_wires:
         try:
-            rotations, diag_obs = qml.grouping.diagonalize_qwc_pauli_words(tape._obs_sharing_wires)
+            rotations, diag_obs = qml.groups.diagonalize_qwc_pauli_words(tape._obs_sharing_wires)
         except ValueError as e:
             raise qml.QuantumFunctionError(
                 "Only observables that are qubit-wise commuting "
